backend2/requirement.txt
Flask
Flask-SQLAlchemy
flask-login
flask-cors
psycopg2-binary
sqlalchemy
sqlalchemy-cockroachdb
